tests(logging): move tests to file
