+++ common/src/io/mod.rs
@@ -1,3 +1,2 @@
 pub mod file;
 pub mod logging;
-mod tests;
